+++ graphql/resolvers/type.go
@@ -8,7 +8,7 @@ import (
 )
 
 func (r *queryResolver) TypeByTypeID(ctx context.Context, id int) (*neo.Type, error) {
-	return r.Services.Universe.Type(ctx, uint64(id))
+	return r.Services.Universe.Type(ctx, uint(id))
 }
 
 func (r *Resolver) Type() service.TypeResolver {
